+++ packages/docusaurus-plugin-content-blog/package.json
@@ -27,7 +27,6 @@
     "feed": "^4.2.2",
     "fs-extra": "^10.0.0",
     "globby": "^11.0.2",
-    "js-yaml": "^4.0.0",
     "loader-utils": "^2.0.0",
     "lodash": "^4.17.20",
     "reading-time": "^1.5.0",
